requirements.txt
africastalking==1.2.7
asgiref==3.7.2
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
contextlib2==21.6.0
coverage==7.4.1
cryptography==42.0.2
defusedxml==0.7.1
distlib==0.3.8
dj-rest-auth==5.0.2
Django==5.0.2
django-allauth==0.57.2
django-debug-toolbar==4.3.0
django-phonenumber-field==7.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-nested-routers==0.93.5
drf-spectacular==0.27.1
filelock==3.13.1
identify==2.5.34
idna==3.6
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
nodeenv==1.8.0
oauthlib==3.2.2
phonenumbers==8.13.30
pillow==10.2.0
platformdirs==4.2.0
pre-commit==3.6.1
pycparser==2.21
PyJWT==2.8.0
python-decouple==3.8
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.18.0
schema==0.7.5
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.0
virtualenv==20.25.0
